LUN-500: Replace RNModal with inbuild navigation modal
