+++ src/navigation/NavigationTypes.ts
@@ -79,6 +79,7 @@ export type RoutesParams = {
   Imprint: undefined
   ManageSelection: undefined
   Settings: undefined
+  OverlayMenu: undefined
   Sponsors: undefined
   Favorites: undefined
   Dictionary: undefined
